# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API каталога исполнителей и их альбомов с песнями",
        "description": "API каталога исполнителей и их альбомов с песнями",
        "version": "0.0.1"
    },
    "paths": {
...
        "/api/executors/": {
            "get": {
                "tags": [
                    "Исполнители"
                ],
                "summary": "Получение информации об исполнителе",
                "description": "Получение массива всех исполнителей",
                "operationId": "GetExecutors",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ExecutorResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Исполнители"
                ],
                "summary": "Добавление исполнителя",
                "description": "Добавление исполнителя",
                "operationId": "PostExecutor",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExecutorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ExecutorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/executors/{executor_id}": {
            "delete": {
                "tags": [
                    "Исполнители"
                ],
                "summary": "Удаление исполнителя",
                "description": "Удаление исполнителя и всей информации о нем",
                "operationId": "DeletehExecutor",
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Исполнитель не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Entity Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Исполнители"
                ],
                "summary": "Обновление информации об исполнителе",
                "description": "Обновление информации об исполнителе",
                "operationId": "PatchExecutor",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExecutorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ExecutorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Исполнитель не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Entity Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при выполнении запроса",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке",
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ExecutorRequest": {
                "description": "Данные запроса для добавления исполнителя",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Название исполнителя",
                        "type": "string",
                        "example": "Yegor Letov"
                    }
                },
                "type": "object"
            },
...
            "ExecutorResource": {
                "description": "Данные об исполнителе",
                "properties": {
                    "id": {
                        "description": "Id исполнителя",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Название исполнителя",
                        "type": "string",
                        "example": "Corey Taylor"
                    }
                },
                "type": "object"
            },
...
        },
...
    }
}